complete DAO,Service and spring bean config
