+++ src/main/java/com/dao/ArticleDao.java
@@ -1,21 +1,17 @@
-package com.service;
-
-import com.bean.Paragraph;
+package com.dao;
 
 import java.util.List;
 
 /**
- * Created by yjy on 16-10-24.
+ * Created by yjy on 16-10-26.
  */
-public interface ArticleService {
+public interface ArticleDao {
 
     @SuppressWarnings("rawtypes")
     public List find_article_of_journal(Integer journal_id);
 
     @SuppressWarnings("rawtypes")
     public List find_chapter_of_article(Integer article_id);
-
-    public List<List<Paragraph>> find_paragraph_of_article(Integer chapter_id);
 
     public void update_article(Integer article_id, String article_title, String article_outline);
 }
