+++ src/main/java/com/dao/UserDAO.java
@@ -2,6 +2,8 @@ package com.dao;
 
 import com.bean.User;
 
+import java.util.List;
+
 public interface UserDAO {
 	public void saveUser(User user);
 	
@@ -13,4 +15,6 @@ public interface UserDAO {
 
 	public String find_note_of_user(Integer userid);
 
+	@SuppressWarnings("rawtypes")
+	public List find_all_paragraph_of_user(Integer userid);
 }
